+++ src/main/java/io/catalyte/training/sportsproducts/domains/review/Review.java
@@ -16,7 +16,7 @@ public class Review {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
   private String title;
-  private int rating;
+  private Double rating;
   private String review;
   private String createdAt;
   private String userName;
@@ -25,7 +25,7 @@ public class Review {
   @JsonIgnore
   private Product product;
   public Review(){}
-  public Review(String title, int rating, String review, String createdAt, String userName, String userEmail, Product product) {
+  public Review(String title, Double rating, String review, String createdAt, String userName, String userEmail, Product product) {
     this.title = title;
     this.rating = rating;
     this.review = review;
@@ -75,11 +75,11 @@ public class Review {
     this.userEmail = userEmail;
   }
 
-  public int getRating() {
+  public Double getRating() {
     return rating;
   }
 
-  public void setRating(int rating) {
+  public void setRating(Double rating) {
     this.rating = rating;
   }
